requirements.txt
anyio==4.10.0
base58==2.1.1
blinker==1.9.0
cachelib==0.9.0
certifi==2025.8.3
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.2.18
Flask==3.1.0
Flask-Caching==2.1.0
Flask-Limiter==3.10.1
Flask-WTF==1.2.2
fonttools==4.59.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.9
limits==4.0.1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
numpy==2.3.2
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
upstash-redis==1.4.0
urllib3==2.3.0
user-agents==2.2.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
WTForms==3.2.1
